// src/sdk/models/DocumentSummaryEntity.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DocumentSummaryEntity = {
  /**
   * Document ID
   */
  id: string;
  /**
   * Document name
   */
  name: string;
  /**
   * Document path
   */
  path: string;
};